Initial cleanup of admin settings code + consolidate map settings
