+++ frontend/src/metabase/admin/settings/components/SettingHeader.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const SettingHeader = ({ setting }) =>
+    <div>
+        <div className="text-grey-4 text-bold text-uppercase">{setting.display_name}</div>
+        <div className="text-grey-4 my1">
+            {setting.description}
+            {setting.note && <div>{setting.note}</div>}
+        </div>
+    </div>
+
+export default SettingHeader;
