+++ frontend/src/metabase/admin/settings/components/widgets/SettingPassword.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+import SettingInput from "./SettingInput.jsx";
+
+const SettingPassword = (props) =>
+    <SettingInput {...props} type="password" />;
+
+export default SettingPassword;
