+++ frontend/src/metabase/admin/settings/selectors.js
@@ -4,6 +4,8 @@ import MetabaseSettings from "metabase/lib/settings";
 
 import { slugify } from "metabase/lib/formatting";
 
+import CustomGeoJSONWidget from "./components/widgets/CustomGeoJSONWidget.jsx";
+
 const SECTIONS = [
     {
         name: "Setup",
@@ -47,11 +49,6 @@ const SECTIONS = [
                 key: "enable-advanced-humanization",
                 display_name: "Friendly Table and Field Names",
                 type: "boolean"
-            },
-            {
-                key: "map-tile-server-url",
-                display_name: "Map tile server URL",
-                type: "string"
             }
         ]
     },
@@ -151,10 +148,19 @@ const SECTIONS = [
         ]
     },
     {
-        name: "Custom Maps",
+        name: "Maps",
         settings: [
             {
-                key: "custom-geojson"
+                key: "map-tile-server-url",
+                display_name: "Map tile server URL",
+                type: "string"
+            },
+            {
+                key: "custom-geojson",
+                display_name: "Custom Maps",
+                description: "Add your own GeoJSON files to enable different region map visualizations",
+                widget: CustomGeoJSONWidget,
+                noHeader: true
             }
         ]
     }
